+++ plugins/woocommerce-admin/packages/experimental/CHANGELOG.md
@@ -1,8 +1,11 @@
-# Unreleased
+# 1.4.0
 
 -   Add new VerticalCSSTransition component for handling height transitions. #7203
--   Update TaskItem to make use of the VerticalCSSTransition. #7203
+-   Add a delete option to completed tasks #7300
 -   Make the action button optionable in TaskItem. #7263
+-   Fix WordPress 5.8 compatibility UI fixes #7255
+-   Fix inbox note dismiss dropdown not closing on Safari #7278
+-   Update TaskItem to make use of the VerticalCSSTransition. #7203
 -   Update CollapsibleList to support nested transitional items. #7263
 
 # 1.3.0